langchain-langserve-apps/streamlit-app/requirements.txt
arize
arize[LLM_Evaluation]
arize-otel
arize-phoenix
arize-phoenix-evals
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-langchain
openinference-instrumentation-dspy
python-dotenv
fastapi
firecrawl-py
gcloud
guardrails-ai
html2text
huggingface_hub
ipython
instructor
jupyter
langchain==0.2.5
langchain-chroma
langchain_cohere
langchain-community
langchain_experimental
langchain-fireworks
langchain-google-genai
langchain-groq
langgraph
langchain_google_community
langchain-mistralai
langchain-openai
langchain-pinecone
langchain-together
langserve[all]
langsmith
numpy
openinference-instrumentation-guardrails
pandas
pinecone
pydantic
pypandoc
streamlit
tavily-python
text-generation
transformers
wandb
unstructured
unstructured[pdf]

